Refactor GraqlShell to be more extensible
