+++ grakn-bootup/src/main/java/ai/grakn/bootup/Graql.java
@@ -18,13 +18,19 @@
 
 package ai.grakn.bootup;
 
-import ai.grakn.graql.shell.GraqlShell;
+import ai.grakn.engine.GraknConfig;
+import ai.grakn.graql.shell.GraknSessionProvider;
+import ai.grakn.graql.shell.GraqlConsole;
+import ai.grakn.graql.shell.GraqlShellOptions;
+import ai.grakn.graql.shell.SessionProvider;
 import ai.grakn.migration.csv.CSVMigrator;
 import ai.grakn.migration.export.Main;
 import ai.grakn.migration.json.JsonMigrator;
 import ai.grakn.migration.sql.SQLMigrator;
 import ai.grakn.migration.xml.XmlMigrator;
 import ai.grakn.util.GraknVersion;
+import com.google.common.base.StandardSystemProperty;
+import org.apache.commons.cli.ParseException;
 
 import java.io.IOException;
 import java.util.Arrays;
@@ -35,6 +41,14 @@ import java.util.Arrays;
  */
 public class Graql {
 
+    private SessionProvider sessionProvider;
+    private static final String HISTORY_FILENAME = StandardSystemProperty.USER_HOME.value() + "/.graql-history";
+
+
+    public Graql(SessionProvider sessionProvider) {
+        this.sessionProvider = sessionProvider;
+    }
+
     /**
      *
      * Invocation from bash script 'graql'
@@ -42,7 +56,7 @@ public class Graql {
      * @param args
      */
     public static void main(String[] args) throws IOException, InterruptedException {
-        new Graql().run(args);
+        new Graql(new GraknSessionProvider(GraknConfig.create())).run(args);
     }
 
     public void run(String[] args) throws IOException, InterruptedException {
@@ -50,7 +64,16 @@ public class Graql {
 
         switch (context) {
             case "console":
-                GraqlShell.main(valuesFrom(args, 1));
+                GraqlShellOptions options;
+
+                try {
+                    options = GraqlShellOptions.create(valuesFrom(args, 1));
+                } catch (ParseException e) {
+                    System.err.println(e.getMessage());
+                    return;
+                }
+
+                GraqlConsole.start(options, sessionProvider, HISTORY_FILENAME);
                 break;
             case "migrate":
                 migrate(valuesFrom(args, 1));
