+++ grakn-graql-shell/src/main/java/ai/grakn/graql/shell/GraqlShell.java
@@ -106,6 +106,8 @@ public class GraqlShell implements AutoCloseable {
     private static final String LICENSE_COMMAND = "license";
     private static final String CLEAN_COMMAND = "clean";
 
+    private boolean errorOccurred = false;
+
     /**
      * Array of available commands in shell
      */
@@ -113,8 +115,6 @@ public class GraqlShell implements AutoCloseable {
             EDIT_COMMAND, COMMIT_COMMAND, ROLLBACK_COMMAND, LOAD_COMMAND, DISPLAY_COMMAND, CLEAR_COMMAND, EXIT_COMMAND,
             LICENSE_COMMAND, CLEAN_COMMAND
     );
-
-    private static final String HISTORY_FILENAME = StandardSystemProperty.USER_HOME.value() + "/.graql-history";
 
     private final OutputFormat outputFormat;
     private final boolean infer;
@@ -129,111 +129,9 @@ public class GraqlShell implements AutoCloseable {
     private final GraqlCompleter graqlCompleter;
     private final ExternalEditor editor = ExternalEditor.create();
 
-    private boolean errorOccurred = false;
 
-    /**
-     * Run a Graql REPL
-     *
-     * @param args arguments to the Graql shell. Possible arguments can be listed by running {@code graql console --help}
-     */
-    public static void main(String[] args) throws InterruptedException, IOException {
-        boolean success = runShell(args, GraknVersion.VERSION, HISTORY_FILENAME, GraknConfig.create());
-        System.exit(success ? 0 : 1);
-    }
 
-    public static boolean runShell(
-            String[] args, String version, String historyFilename, GraknConfig config
-    ) throws InterruptedException, IOException {
-
-        GraqlShellOptions options;
-
-        try {
-            options = GraqlShellOptions.create(args);
-        } catch (ParseException e) {
-            System.err.println(e.getMessage());
-            return false;
-        }
-
-        List<String> queries = null;
-
-        String query = options.getQuery();
-
-        // This is a best-effort guess as to whether the user has made a mistake, without parsing the query
-        if (query != null) {
-            queries = ImmutableList.of(query);
-
-            if (!query.contains("$") && query.trim().startsWith("match")) {
-                System.err.println(ErrorMessage.NO_VARIABLE_IN_QUERY.getMessage());
-            }
-        }
-
-        List<Path> filePaths = options.getFiles();
-
-        // Print usage message if requested or if invalid arguments provided
-        if (options.displayHelp()) {
-            GraqlShellOptions.printUsage();
-            return true;
-        }
-
-        if (options.displayVersion()) {
-            System.out.println(version);
-            return true;
-        }
-
-        Keyspace keyspace = options.getKeyspace();
-
-        int defaultGrpcPort = config.getProperty(GraknConfigKey.GRPC_PORT);
-        SimpleURI defaultGrpcUri = new SimpleURI(Grakn.DEFAULT_URI.getHost(), defaultGrpcPort);
-
-        SimpleURI location = options.getUri();
-        OutputFormat outputFormat = options.getOutputFormat();
-
-        SimpleURI httpUri = location != null ? location : Grakn.DEFAULT_URI;
-        SimpleURI grpcUri = location != null ? location : defaultGrpcUri;
-
-        boolean infer = options.shouldInfer();
-
-        Path path = options.getBatchLoadPath();
-
-        if (path != null) {
-            try {
-                BatchLoader.sendBatchRequest(httpUri, keyspace, path);
-            } catch (Exception e) {
-                System.out.println("Batch failed \n" + CommonUtil.simplifyExceptionMessage(e));
-                return false;
-            }
-            return true;
-        }
-
-        try (GraqlShell shell = new GraqlShell(historyFilename, keyspace, grpcUri, outputFormat, infer)) {
-            if (filePaths != null) {
-                queries = loadQueries(filePaths);
-            }
-
-            // Start shell
-            shell.start(queries);
-            return !shell.errorOccurred;
-        } catch (StatusRuntimeException e) {
-            if (e.getStatus().getCode().equals(Status.Code.UNAVAILABLE)) {
-                System.err.println(ErrorMessage.COULD_NOT_CONNECT.getMessage());
-                return false;
-            } else {
-                throw e;
-            }
-        }
-    }
-
-    private static List<String> loadQueries(Iterable<Path> filePaths) throws IOException {
-        List<String> queries = Lists.newArrayList();
-
-        for (Path filePath : filePaths) {
-            queries.add(loadQuery(filePath));
-        }
-
-        return queries;
-    }
-
-    private static String loadQuery(Path filePath) throws IOException {
+    public static String loadQuery(Path filePath) throws IOException {
         List<String> lines = Files.readAllLines(filePath, StandardCharsets.UTF_8);
         return lines.stream().collect(joining("\n"));
     }
@@ -241,21 +139,19 @@ public class GraqlShell implements AutoCloseable {
     /**
      * Create a new Graql shell
      */
-    private GraqlShell(
-            String historyFilename, Keyspace keyspace, SimpleURI uri, OutputFormat outputFormat, boolean infer
-    ) throws IOException {
+    public GraqlShell(String historyFilename, GraknSession session, ConsoleReader console, OutputFormat outputFormat, boolean infer) throws IOException {
         this.outputFormat = outputFormat;
         this.infer = infer;
-        console = new ConsoleReader(System.in, System.out);
+        this.console = console;
+        this.session = session;
+        this.graqlCompleter = GraqlCompleter.create(session);
+        this.historyFile = HistoryFile.create(console, historyFilename);
 
-        session = RemoteGrakn.session(uri, keyspace);
+
         tx = session.open(GraknTxType.WRITE);
-
-        graqlCompleter = GraqlCompleter.create(session);
-        historyFile = HistoryFile.create(console, historyFilename);
     }
 
-    private void start(@Nullable List<String> queryStrings) throws IOException, InterruptedException {
+    public void start(@Nullable List<String> queryStrings) throws IOException, InterruptedException {
         try {
             if (queryStrings != null) {
                 for (String queryString : queryStrings) {
@@ -422,6 +318,10 @@ public class GraqlShell implements AutoCloseable {
     private void reopenTx() {
         if (!tx.isClosed()) tx.close();
         tx = session.open(GraknTxType.WRITE);
+    }
+
+    public boolean errorOccurred() {
+        return errorOccurred;
     }
 
     @Override
